RALPH: Review - extract shared session-cookie helper, drop unused import
The two new operator-authenticated routes added in this slice (agents,
conversations) carried identical copies of `SESSION_COOKIE`,
`resolveOrgId`, and `hashToken`. Pulled them into
`routes/_session-cookie.ts` so each route file just imports the
resolver. `auth.ts` keeps its own copy because session issuance also
lives there and isn't part of this slice.

Also dropped an unused `agentMessage` import in
`headless/session-store.test.ts`.

No behaviour change.
