+++ apps/server/src/routes/agents.ts
@@ -1,11 +1,9 @@
-import { Hono, type Context } from 'hono';
-import { getCookie } from 'hono/cookie';
+import { Hono } from 'hono';
 import * as v from 'valibot';
 import { and, eq } from 'drizzle-orm';
 import type { PostgresJsDatabase } from 'drizzle-orm/postgres-js';
-import { agent, channel, session, user } from '../db/schema.ts';
-
-const SESSION_COOKIE = 'nexus_session';
+import { agent, channel } from '../db/schema.ts';
+import { resolveOrgId } from './_session-cookie.ts';
 
 // Headless is the only mode the UI exposes at this slice. The schema accepts
 // `dedicated` so a later slice can flip it on without a migration.
@@ -116,25 +114,6 @@ function toApi(row: AgentRow, widgetChannelId: string | null) {
     voiceEnabled: row.voiceEnabled,
     widgetChannelId,
   };
-}
-
-async function resolveOrgId(c: Context, db: PostgresJsDatabase): Promise<string | null> {
-  const token = getCookie(c, SESSION_COOKIE);
-  if (!token) return null;
-  const [row] = await db
-    .select({ orgId: user.orgId, expiresAt: session.expiresAt })
-    .from(session)
-    .innerJoin(user, eq(session.userId, user.id))
-    .where(eq(session.tokenHash, hashToken(token)))
-    .limit(1);
-  if (!row || row.expiresAt.getTime() < Date.now()) return null;
-  return row.orgId;
-}
-
-function hashToken(token: string): string {
-  const hasher = new Bun.CryptoHasher('sha256');
-  hasher.update(token);
-  return hasher.digest('hex');
 }
 
 async function safeJson(req: Request): Promise<unknown> {
